fix(factory): keep revealing a column while its sentinel stays in view
An IntersectionObserver only reports crossings, so a sentinel that entered
the reveal margin and never left reported once and the column stalled two
pages in — the rest of its cards were never rendered. The sentinel is keyed
by what is revealed, so each step observes it again and a lane short enough
to sit inside the margin renders in full.

The candidate card's detail panel moves into its own component, matching the
work-item card. Its custom-prompt state goes with it: nothing outside the
panel ever read it.
