+++ mastracode/factory-ui/src/ui/domains/factory/components/CandidateCard.tsx
@@ -1,11 +1,9 @@
 import { Button } from '@mastra/playground-ui/components/Button';
 import { DropdownMenu } from '@mastra/playground-ui/components/DropdownMenu';
-import { Popover, PopoverContent } from '@mastra/playground-ui/components/Popover';
-import { Textarea } from '@mastra/playground-ui/components/Textarea';
 import { cn } from '@mastra/playground-ui/utils/cn';
-import { ArrowUpRight, EllipsisVertical, Minimize2, PencilLine, Plus } from 'lucide-react';
+import { ArrowUpRight, EllipsisVertical, Plus } from 'lucide-react';
 import type { ReactElement } from 'react';
-import { useId, useRef, useState } from 'react';
+import { useId } from 'react';
 
 import { useCardMorph } from '../hooks/useCardMorph';
 import type { FactoryRunPhase } from '../../../../hooks/useStartFactoryRun';
@@ -22,9 +20,8 @@ import {
   REVEAL_ON_CARD_HOVER,
   SourceTitle,
 } from './BoardCardParts';
-import { CardSourceDescription } from './BoardCardDetails';
 import { SourceIcon, actionIcon } from './BoardIcons';
-import { CardDetailsBody, CardDetailsPanel } from './CardDetailsPanel';
+import { CandidateDetailsPanel } from './CandidateDetailsPanel';
 
 /**
  * A GitHub/Linear item with no work-item record yet. Acting on it is what
@@ -55,9 +52,6 @@ export function CandidateCard({
   onFile: () => void;
 }) {
   const detailsTitleId = useId();
-  const promptAnchorRef = useRef<HTMLButtonElement>(null);
-  const [promptOpen, setPromptOpen] = useState(false);
-  const [prompt, setPrompt] = useState('');
   const morph = useCardMorph();
 
   const labels = metadataLabels(candidate.metadata);
@@ -69,24 +63,6 @@ export function CandidateCard({
       .map(action => ({ label: action.label, phase: pendingRunRoles.get(action.role) })),
     preparing,
   });
-
-  const closePrompt = () => {
-    setPromptOpen(false);
-    setPrompt('');
-  };
-
-  const runPrompt = () => {
-    const trimmed = prompt.trim();
-    if (!trimmed || runPending) return;
-    closePrompt();
-    morph.closeDetails();
-    onRun(defaultAction, trimmed);
-  };
-
-  const startDefaultRun = () => {
-    morph.closeDetails();
-    onRun(defaultAction);
-  };
 
   const fileFromDetails = () => {
     morph.closeDetails();
@@ -197,118 +173,19 @@ export function CandidateCard({
         </article>
       </CardTitleTooltip>
 
-      <CardDetailsPanel morph={morph} labelledBy={detailsTitleId}>
-        {/* The card's own rows, in the card's own order, spacing and padding:
-            what the card already showed is where it already was, so the box
-            grows around it instead of re-staging it. */}
-        <div className="flex flex-col gap-3 p-3">
-          <div className="flex min-w-0 flex-col gap-1.5">
-            <div className="flex min-w-0 items-center gap-1.5">
-              <span className="text-ui-xs text-icon2 min-w-0 flex-1 truncate">{candidate.meta}</span>
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon-xs"
-                aria-label={`Collapse ${candidate.title}`}
-                onClick={morph.closeDetails}
-              >
-                <Minimize2 size={13} aria-hidden />
-              </Button>
-              <DropdownMenu>
-                <DropdownMenu.Trigger
-                  render={
-                    <Button
-                      type="button"
-                      variant="ghost"
-                      size="icon-xs"
-                      aria-label={`All actions for ${candidate.title}`}
-                    >
-                      <EllipsisVertical size={13} aria-hidden />
-                    </Button>
-                  }
-                />
-                <DropdownMenu.Content align="end" className="min-w-44">
-                  {menuItems}
-                </DropdownMenu.Content>
-              </DropdownMenu>
-            </div>
-            <div className="flex min-w-0 items-center gap-1.5">
-              <SourceIcon source={candidate.source} />
-              <h2 id={detailsTitleId} className="text-ui-smd text-icon6 m-0 min-w-0 font-semibold wrap-anywhere">
-                {candidate.title}
-              </h2>
-            </div>
-          </div>
-          <CardLabels labels={labels} />
-        </div>
-        {/* Only what the card never carried is staged in. */}
-        <CardDetailsBody>
-          <CardSourceDescription
-            item={candidate}
-            projectRepositoryId={projectRepositoryId}
-            factoryProjectId={factoryProjectId}
-          />
-        </CardDetailsBody>
-        <div className="flex flex-col gap-2 px-3 py-2.5" data-card-morph="reveal">
-          <Button
-            ref={promptAnchorRef}
-            type="button"
-            variant="outline"
-            size="sm"
-            className="w-full"
-            onClick={() => setPromptOpen(true)}
-          >
-            <PencilLine size={13} aria-hidden />
-            Custom prompt…
-          </Button>
-          <Button
-            type="button"
-            variant="primary"
-            size="sm"
-            className="w-full"
-            disabled={disabled || runPending}
-            onClick={startDefaultRun}
-          >
-            {defaultAction.label}
-          </Button>
-          <Popover open={promptOpen} onOpenChange={open => (open ? setPromptOpen(true) : closePrompt())}>
-            <PopoverContent anchor={promptAnchorRef} align="end" className="w-80 p-3">
-              <form
-                aria-label={`Custom prompt for ${candidate.title}`}
-                className="flex flex-col gap-2"
-                onSubmit={event => {
-                  event.preventDefault();
-                  runPrompt();
-                }}
-              >
-                <Textarea
-                  autoFocus
-                  rows={3}
-                  size="sm"
-                  value={prompt}
-                  placeholder="What should the agent do with this?"
-                  aria-label={`Prompt for ${candidate.title}`}
-                  onChange={event => setPrompt(event.target.value)}
-                  onKeyDown={event => {
-                    if (event.key === 'Enter' && (event.metaKey || event.ctrlKey)) {
-                      event.preventDefault();
-                      runPrompt();
-                    }
-                  }}
-                />
-                <div className="flex justify-end gap-2">
-                  <Button type="button" variant="ghost" size="xs" onClick={closePrompt}>
-                    Cancel
-                  </Button>
-                  <Button type="submit" size="xs" disabled={runPending || !prompt.trim()}>
-                    Run
-                  </Button>
-                </div>
-              </form>
-            </PopoverContent>
-          </Popover>
-        </div>
-      </CardDetailsPanel>
+      <CandidateDetailsPanel
+        candidate={candidate}
+        labelledBy={detailsTitleId}
+        morph={morph}
+        labels={labels}
+        projectRepositoryId={projectRepositoryId}
+        factoryProjectId={factoryProjectId}
+        menu={menuItems}
+        defaultAction={defaultAction}
+        disabled={disabled}
+        runPending={runPending}
+        onRun={onRun}
+      />
     </>
   );
 }
