+++ mastracode/factory-ui/src/ui/domains/factory/components/ColumnReveal.tsx
@@ -26,8 +26,10 @@ export function ColumnReveal<T>({
   const pinnedIndex = pinned === undefined ? -1 : items.findIndex(pinned);
   const count = Math.max(revealed, pinnedIndex + 1);
 
-  // Identity must stay stable: a fresh observer on a sentinel that is already
-  // in view fires at once, and each firing would reveal another step.
+  // Keyed by what is revealed, so each step remounts the sentinel and observes
+  // it again. An observer only reports crossings: a sentinel that never leaves
+  // view — a lane whose cards all fit — would otherwise report once and stall
+  // with the rest of the column unrendered.
   const sentinelRef = useCallback((node: HTMLDivElement | null) => {
     if (node === null) return;
     const observer = new IntersectionObserver(
@@ -43,7 +45,7 @@ export function ColumnReveal<T>({
   return (
     <>
       {items.slice(0, count).map(renderItem)}
-      {count < items.length && <div ref={sentinelRef} aria-hidden className="h-px shrink-0" />}
+      {count < items.length && <div key={count} ref={sentinelRef} aria-hidden className="h-px shrink-0" />}
     </>
   );
 }
